Floundering
Oh, that I had had more rigorous training in type theory! And oh, that I
could make the time now to take it up! I fear that this project does not
look as promising as it once did.

Perhaps I have overcomplicated and overgeneralized. Also it is hard to
get one's brain really wrapped around such things in the evening.